+++ api/CHANGELOG.md
@@ -14,6 +14,8 @@
 - Added a search button next to all version listings under "Show suggested versions" on the app selection page. Tapping the new button searches with google the package name and the version number (example: com.google.android.youtube 20.51.38)
 - Redesigned the UI of the app selection page
 - Patch filter selections on the patch selection page now persist https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/73
+- Improved patch selection screen UI
+- Added chip tags on patches on the patch selection screen showing the versions said patch supports https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/75
 
 
 # Bug fixes